relocate number formatting function fix dark mode select input color
