+++ styles/globals.scss
@@ -187,3 +187,7 @@ rect.selection {
   background-color: #A463B0;
   color: white;
 }
+
+.react_select__input {
+  color: $text !important
+}
